Change all routes to remove epjs naming #1 #6
