+++ test/controllers/address_controller_test.exs
@@ -16,17 +16,17 @@ defmodule EpjsApp.AddressControllerTest do
   end
 
   describe "address controller returns json response with valid slam_id" do
-    test "GET /epjspatientaddressdetails/address/get-address?service_user", %{conn: conn} do
+    test "GET /patient-address-details/address/get-address?service_user", %{conn: conn} do
       stringified_user = %{slam_id: 123} |> Poison.encode!
-      conn = get conn, "http://localhost:4001/epjspatientaddressdetails/address/get-address?service_user=" <> stringified_user
+      conn = get conn, "http://localhost:4001/patient-address-details/address/get-address?service_user=" <> stringified_user
       assert json_response(conn, 200)
     end
   end
 
   describe "address controller returns null json response with invalid slam_id" do
-    test "GET /epjspatientaddressdetails/address/get-address?service_user", %{conn: conn} do
+    test "GET /patient-address-details/address/get-address?service_user", %{conn: conn} do
       stringified_user = %{slam_id: 456} |> Poison.encode!
-      conn = get conn, "http://localhost:4001/epjspatientaddressdetails/address/get-address?service_user=" <> stringified_user
+      conn = get conn, "http://localhost:4001/patient-address-details/address/get-address?service_user=" <> stringified_user
       assert conn.resp_body == "null"
     end
   end
