+++ web/router.ex
@@ -8,11 +8,11 @@ defmodule EpjsApp.Router do
   scope "/", EpjsApp do
     pipe_through :api
 
-    get "/care-team/for", CareTeamController, :for
     get "/user/carer-connection/find-user", UserController, :find_user
-    get "/epjsteammember/clinician-connection/find-clinician", TeamMemberController, :find_clinician
-    get "/epjsteammember/clinician-connection/get-patients", TeamMemberController, :get_patients
-    get "/epjspatientaddressdetails/address/get-address", AddressController, :get_address
+    get "/team-member/care-team/for", TeamMemberController, :for
+    get "/team-member/clinician-connection/find-clinician", TeamMemberController, :find_clinician
+    get "/team-member/clinician-connection/get-patients", TeamMemberController, :get_patients
+    get "/patient-address-details/address/get-address", AddressController, :get_address
 
   end
 end
